fix DB schema for relation integrity, add FA icon picker for admin

+++ database/migrations/2019_12_10_064519_create_page_contents_table.php
@@ -15,8 +15,8 @@ class CreatePageContentsTable extends Migration
     {
         Schema::create('page_contents', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('page_id');
-            $table->integer('user_id');
+            $table->bigInteger('page_id');
+            $table->bigInteger('user_id');
 
             $table->string('title');
             $table->binary('content');
